% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ah.R
\name{ah}
\alias{ah}
\alias{nah}
\title{Paste Wrappers - Preset Collapse}
\usage{
ah(x, collapse = " ", ...)

nah(x, collapse = "\\n", ...)
}
\arguments{
\item{x}{A vector to collapse.}

\item{collapse}{An character string to separate the results.}

\item{\ldots}{Other arguments passed to \code{\link[base]{paste}}.}
}
\value{
Returns a collapsed, pasted vector.
}
\description{
\code{ah} - \code{\link[base]{paste}} with \code{collapse = " "}.  The name
comes from the interjection "ah" as in "Ah I don't need to type
\code{collapse = " "}."

\code{nah} - \code{\link[base]{paste}} with \code{collapse = "\n"}, useful
in concert with \code{\link[base]{cat}}.  The name is simpley \code{ah} +
\code{n} from \code{collapse = "\n"}.
}
\examples{
ah(LETTERS[1:8])
cat(nah(LETTERS[1:8]))
}

