% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/highlight_style.R
\name{highlight_style}
\alias{highlight_style}
\title{Generate an HTML/CSS Style}
\usage{
highlight_style(marker_color)
}
\arguments{
\item{marker_color}{A named list or vector of colors.  The colors will be used
for highlighting and the names will correspond to the class name of the marker
tag.}
}
\value{
Returns a vector with an HTML/CSS style template.
}
\description{
Generate an HTML/CSS style syntax for highlight marker tags.  Useful in
\pkg{knitr}/\pkg{rmarkdown} style documents.
}
\examples{
marks <- c(pos = "lightgreen", neg = "pink", neutral = "yellow")
highlight_style(marks)

## A .Rmd Example
## 1. Read the lines below into `content`
## 2. Paste together and `cat` a file out
## 3. `render` via rmarkdown package
## 4. Open the html document
\dontrun{
content <- readLines(n=26)
---
title: "Test"
date: "`r format(Sys.time(), '\%d \%B, \%Y')`"
output:
  html_document:
    toc: true
    theme: journal
    number_sections: true
---

```{r, echo=FALSE}
pacman::p_load(hilight)
x <-"I like good cheeses but she hated Bad cheese"
stems <- c("cheese", "like", "good", "bad")
marks_class <- c("neutral", "pos", "pos", "neg")
marks <- c(pos = "lightgreen", neg = "pink", neutral = "yellow")
body <- mark_word_stems(x, stems, marks_class)
```

```{r, echo=FALSE, results='asis'}
highlight_style(marks)
```

```{r, echo=FALSE, results='asis'}
cat(body)
```

cat(paste(content, collapse="\\n"), "\\n", file="test.Rmd")
rmarkdown::render("test.Rmd")
browseURL("test.html")
}
}

