% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight_template.R
\name{highlight_template}
\alias{highlight_template}
\title{Generate an HTML Highlight Template}
\usage{
highlight_template(marker_color)
}
\arguments{
\item{marker_color}{A named list or vector of colors.  The colors will be used
for highlighting and the names will correspond to the class name of the marker
tag.}
}
\value{
Returns a vector with an HTML read template.
}
\description{
Generate an HTML syntax with css for highlight marker tags that takes a body
insert and can then be used to generate an HTML output with text highlighting.
}
\examples{
marks <- c(pos = "lightgreen", neg = "pink", neutral = "yellow")
temp <- highlight_template(marks)
body <- paste0("I like <mark class=\\"pos\\">good</mark>",
    " <mark class=\\"neutral\\">fish</mark>",
    ", but <mark class=\\"neg\\">bad</mark> <mark class=\\"neutral\\">fish</mark>",
    " is <mark class=\\"neg\\">gross</mark>.")

template2html(insert_body(highlight_template(marks), body), file="")

\dontrun{
## To an external file
template2html(insert_body(highlight_template(marks), body))
open_html()
}
}

