% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mark_regex.R
\name{mark_regex}
\alias{mark_regex}
\title{Mark Words Containing Regex With Class for Highlighting}
\usage{
mark_regex(x, regex, marks_class, ignore.case = FALSE)
}
\arguments{
\item{x}{A character vector with regex to operate on.}

\item{regex}{The regex to search for.}

\item{marks_class}{The classes to apply to word (must be equal to regex in
length).  These are dictated by the css generated for marks in
\code{highlight_template}.}

\item{ignore.case}{logical.  If \code{TRUE} case is ignored.}
}
\value{
Returns a character vector with mark tags added around words that
contain the regex.
}
\description{
Mark words containing regex with a class for highlighting.
}
\examples{
x <-"I like good cheeses but she hated Bad cheese.  Really hates it"
regex <- c("chees[^ ]*\\\\b", "like", "hate[^ ]*\\\\b", "good", "bad\\\\b")
marks_class <- c("neutral", "pos", "neg", "pos", "neg")
marks <- c(pos = "lightgreen", neg = "pink", neutral = "yellow")

(body <- mark_regex(x, regex, marks_class))
\dontrun{
## To an external file
template2html(insert_body(highlight_template(marks), body))
open_html()
}
}
\seealso{
Other mark.functions: \code{\link{mark_sentences}};
  \code{\link{mark_word_stems}}; \code{\link{mark_words}}
}

