% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark_words.R
\name{mark_words}
\alias{mark_words}
\title{Mark Words With Class for Highlighting}
\usage{
mark_words(x, words, marks_class, ignore.case = TRUE, ...)
}
\arguments{
\item{x}{A character vector with words to operate on.}

\item{words}{The words to search for.}

\item{marks_class}{The classes to apply to word (must be equal to words in
length).  These are dictated by the css generated for marks in
\code{highlight_template}.}

\item{ignore.case}{logical.  If \code{TRUE} case is ignored.}

\item{\ldots}{Other arguments passed to \code{\link[base]{gsub}}.}
}
\value{
Returns a character vector with mark tags added around words.
}
\description{
Mark words with a class for highlighting.
}
\examples{
x <-"I like good cheese but hate Bad cheese"
words <- c("cheese", "like", "good", "bad")
marks_class <- c("neutral", "pos", "pos", "neg")
marks <- c(pos = "lightgreen", neg = "pink", neutral = "yellow")

(body <- mark_words(x, words, marks_class))
\dontrun{
## To an external file
template2html(insert_body(highlight_template(marks), body))
open_html()
}

}
\seealso{
Other mark.functions: \code{\link{mark_regex}},
  \code{\link{mark_sentences}},
  \code{\link{mark_word_stems}}
}

