% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_html.R
\name{open_html}
\alias{open_html}
\title{Open an HTML File With Default Browser}
\usage{
open_html(x = "out.html", ...)
}
\arguments{
\item{x}{An HTML file to open.}

\item{\ldots}{Other arguments passed to \code{\link[utils]{browseURL}}.}
}
\description{
Open an HTML file with default browser.
}
\seealso{
\code{\link[utils]{browseURL}}
}
\keyword{open}

