% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/highlight_style.R
\name{print.highlight_style}
\alias{print.highlight_style}
\title{Prints a highlight_style Object}
\usage{
\method{print}{highlight_style}(x, pretty = TRUE, ...)
}
\arguments{
\item{x}{The highlight_style object.}

\item{pretty}{logical.  If \code{TRUE} \code{\link[base]{cat}} will be used
to print the object.}

\item{\ldots}{Other arguments passed to \code{\link[base]{cat}}}
}
\description{
Prints a highlight_style object.
}

