% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_sentences.R
\name{split_sentences}
\alias{split_sentences}
\title{Split Sentences}
\usage{
split_sentences(x, ...)
}
\arguments{
\item{x}{A character vector with sentences.}

\item{\ldots}{Ignored.}
}
\value{
Returns a list of vectors of sentences.
}
\description{
Split sentences from a character vector.
}
\examples{
(x <- paste0(
    "Mr. Brown comes! He says hello. i give him coffee.  i will ",
    "go at 5 p. m. eastern time.  Or somewhere in between!go there"
))
split_sentences(x)
}

