% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tag.R
\name{tag}
\alias{b}
\alias{em}
\alias{h1}
\alias{h2}
\alias{h3}
\alias{h4}
\alias{p}
\alias{span}
\alias{tag}
\alias{u}
\title{Wrapp Text With HTML Tags}
\usage{
tag(x, left, right = gsub("^<", "</", left))

p(x)

h1(x)

h2(x)

h3(x)

h4(x)

span(x, extra = "")

b(x)

em(x)

u(x)
}
\arguments{
\item{x}{A text string.}

\item{left}{A left tag.}

\item{right}{A right tag.}

\item{extra}{Additional arguments to pass to a left tag such as styling.}
}
\value{
Returns a vector of strings wrpped in HTML tags.
}
\description{
Wrap text with select HTML tags.  \code{tag} is generic in comparison to
specific tag types such as \code{p}.
}
\examples{
tag(LETTERS, "<iframe>")

p(LETTERS)
h1(paste0(LETTERS, ": ", em(letters)))
p(paste(b(LETTERS), 1:26, letters))

span(LETTERS)
set.seed(10)
(body <-p(span(LETTERS, extra=sprintf("style=\\"color:\%s\\"", sample(colors(), 26)))))
\dontrun{
template2html(insert_body(highlight_template(c(x="")), body))
open_html()
}
}

