% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/template2html.R
\name{template2html}
\alias{template2html}
\title{Generate an HTML FIle from Finished Template}
\usage{
template2html(template, file = "out.html", protect = TRUE, ...)
}
\arguments{
\item{template}{A template thate has been created using
\code{highlight_template} and \code{insert_body}.}

\item{file}{A file to output to.  Use \code{file = ""} to print to console.}

\item{protect}{logical.  If \code{TRUE} only allows a tempalte from
\code{insert_body}.  Use \code{FALSE} to override this behavior.}

\item{\ldots}{Other arguments passed to \code{\link[base]{cat}}.}
}
\value{
Returns the HTML vector that was used to produce \code{file}.
}
\description{
Generate an HTML file from a template that has been created using
\code{highlight_template} and \code{insert_body}.
}
\examples{
marks <- c(pos = "lightgreen", neg = "pink", neutral = "yellow")
\dontrun{
template2html(highlight_template(marks), file="", protect=FALSE)
}

temp <- highlight_template(marks)
body <- paste0("I like <mark class=\\"pos\\">good</mark>",
    " <mark class=\\"neutral\\">fish</mark>",
    ", but <mark class=\\"neg\\">bad</mark> <mark class=\\"neutral\\">fish</mark>",
    " is <mark class=\\"neg\\">gross</mark>.")

template2html(insert_body(highlight_template(marks), body), file="")

\dontrun{
## To an external file
template2html(insert_body(highlight_template(marks), body))
open_html()
}
}

